+++ src/components/Interviews.tsx
@@ -37,7 +37,7 @@ import {
   PointOfContactHeaderCells, 
   PointOfContactCells 
 } from './pointOfContactCells';
-import { SourceHeaderCell, SourceCell } from './sourceCell';
+import { MenteeSourceHeaderCell, MenteeSourceCell } from './MenteeSourceCell';
 
 /**
  * @param forCalibration when true, show additional columns in the table and
@@ -60,7 +60,7 @@ export default function Interviews({ interviews, forCalibration }: {
       <Thead><Tr>
         {showStatus && <Th>状态</Th>}
         <PointOfContactHeaderCells />
-        <SourceHeaderCell />
+        <MenteeSourceHeaderCell />
         <Th>候选人</Th>
         <Th>{forCalibration ? "" : "其他"}面试官</Th>
         {forCalibration && <>
@@ -127,7 +127,7 @@ function InterviewRow({ i, forCalibration, showStatus }: {
     {app && app.user &&
       <PointOfContactCells user={app.user} refetch={refetch} />
     }
-    <Td><SourceCell application={app?.application} /></Td>
+    <MenteeSourceCell application={app?.application} />
     <TdLink href={url}><b>{formatUserName(i.interviewee.name)}</b></TdLink>
 
     <TdLink href={url}><Wrap spacing="2">
